requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.4
